Checkpoint.
The approach here is to bite the bullet, and just take out all the stuff
that is preventing sharing in the schema compiler, then try to cope by
back filling algorithms, relocating them or revisiting them so they work
without the need for this backpointer stuff.

Removed all SchemaComponentFactory
Removed backlinks that point from def to ref.
Removed all calculations that use the backlinks.

That disables many optimizations.

daffodil-core does not compile. Many errors. 

Primary areas of work:
* ElementBase - namespace binding minimization
* ChoiceCombinator - for unparsing - computing choice branch map
* Alignment - determining when we have known alignment and known text
alignment is critical to decent optimizations.
* bitOrderChange - determining when we have a bitOrder change is
critical to decent optimization.

Note: alignment could just be put into the runtime without optimization
for parsing. It is unparsing, where alignment regions tend to require
splitting/double buffering that we need the optimizations very much.

bitOrderChange has same impact on unparsing. When bit order changes, we
require a byte boundary. We could just check this and error, but that
begs the question of how we achieve byte alignment when unparsing.

No doubt there is more. This is what the compilation errors are about
now

DAFFODIL-1444