Checkpoint.
7 tests failing in daffodil-core-unittest.

Several are choice unparser related (due to stub in choiceBranchMap)

One is a circular deadlock due to alignment calculations.
(Not unexpected given that we're creating all sorts of alignment
regions due to no optimizing of alignment.)

We have shown that the compiler is no longer creating a combinatorial
explosion of objects. The number of objects is as expected. 

Many tests of compiler "middle" attributes were commented out as
we're not currently computing those middle attributes. Those tests could
be adapted. They do have good use cases for testing in them.