Add structure to forcing of delays.
Need for serialization to force things is gone.

The way things force delays now, is that things that use delays all
define an initialize method, which must be called to insure all
delays are forced.

The forcing is done not by explicitly forcing the delay objects
themselves, but by initializing the the cyclic objects.

This insures we're not in a circular situation because
it explicitly demands the creation of the objects be completed first
before requesting the delays are evaluatable.

NodeInfo object is now self-initializing.

Also fixed a fragile thing around creating and discarding a Sequence object.

DSOM objects can't be created speculatively because just creating one
adds pending requiredEvaluations to the queues as a side-effect.

However, we need to create a Sequence speculatively for one situation which
is the dfdl:hiddenGroupRef property. So a feature was added to
OOLAG to enable turning off the entire requiredEvaluations system for
an object by overriding a member. DummySequence uses this.

DAFFODIL-2326, DAFFODIL-1879, DAFFODIL-2273, DAFFODIL-2526